requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
cattrs==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.3
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.37.0
itemadapter==0.8.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mwparserfromhell==0.6.6
neo4j==5.28.0
nest-asyncio==1.6.0
numpy==2.2.6
ollama==0.4.8
packaging==25.0
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
Protego==0.4.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
queuelib==1.8.0
requests==2.32.3
requests-cache==1.2.1
requests-file==2.1.0
Scrapy==2.8.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tldextract==5.3.0
tomli==2.2.1
tornado==6.5.1
tqdm==4.65.0
traitlets==5.14.3
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
w3lib==2.3.1
wcwidth==0.2.13
zope.interface==7.2
